+++ src/api/core/organizations.rs
@@ -273,7 +273,7 @@ async fn leave_organization(org_id: OrganizationId, headers: OrgMemberHeaders, c
     }
 
     log_event(
-        EventType::OrganizationUserLeft as i32,
+        EventType::OrganizationUserLeft,
         &membership.uuid,
         &org_id,
         &headers.user.uuid,
@@ -331,7 +331,7 @@ async fn post_organization(
     org.save(&conn).await?;
 
     log_event(
-        EventType::OrganizationUpdated as i32,
+        EventType::OrganizationUpdated,
         org_id.as_ref(),
         &org_id,
         &headers.user.uuid,
@@ -655,7 +655,7 @@ async fn post_organization_collections(
     // Emit the success event only after all requested assignments and the creator's object-scoped
     // manage grant have been persisted. A later write failure must not leave a false audit record.
     log_event(
-        EventType::CollectionCreated as i32,
+        EventType::CollectionCreated,
         &collection.uuid,
         &org_id,
         &headers.user.uuid,
@@ -735,7 +735,7 @@ async fn post_bulk_access_collections(
         collection.save(&conn).await?;
 
         log_event(
-            EventType::CollectionUpdated as i32,
+            EventType::CollectionUpdated,
             &collection.uuid,
             &org_id,
             &headers.user.uuid,
@@ -825,7 +825,7 @@ async fn post_organization_collection_update(
     collection.save(&conn).await?;
 
     log_event(
-        EventType::CollectionUpdated as i32,
+        EventType::CollectionUpdated,
         &collection.uuid,
         &org_id,
         &headers.user.uuid,
@@ -895,7 +895,7 @@ async fn delete_organization_collection_impl(
         err!("Collection not found", "Collection does not exist or does not belong to this organization")
     };
     log_event(
-        EventType::CollectionDeleted as i32,
+        EventType::CollectionDeleted,
         &collection.uuid,
         org_id,
         &headers.user.uuid,
@@ -1567,7 +1567,7 @@ async fn send_invite(
         }
 
         log_event(
-            EventType::OrganizationUserInvited as i32,
+            EventType::OrganizationUserInvited,
             &new_member.uuid,
             &org_id,
             &headers.user.uuid,
@@ -1893,7 +1893,7 @@ async fn confirm_invite_impl(
     OrgPolicy::check_user_allowed(&member_to_confirm, "confirm", conn).await?;
 
     log_event(
-        EventType::OrganizationUserConfirmed as i32,
+        EventType::OrganizationUserConfirmed,
         &member_to_confirm.uuid,
         org_id,
         &headers.user.uuid,
@@ -2233,7 +2233,7 @@ async fn edit_member(
     }
 
     log_event(
-        EventType::OrganizationUserUpdated as i32,
+        EventType::OrganizationUserUpdated,
         &member_to_edit.uuid,
         &org_id,
         &headers.user.uuid,
@@ -2320,7 +2320,7 @@ async fn delete_member_impl(
     }
 
     log_event(
-        EventType::OrganizationUserRemoved as i32,
+        EventType::OrganizationUserRemoved,
         &member_to_delete.uuid,
         org_id,
         &headers.user.uuid,
@@ -2799,7 +2799,7 @@ async fn put_policy(
                 }
 
                 log_event(
-                    EventType::OrganizationUserRemoved as i32,
+                    EventType::OrganizationUserRemoved,
                     &member.uuid,
                     &org_id,
                     &headers.user.uuid,
@@ -2825,7 +2825,7 @@ async fn put_policy(
     policy.save(&conn).await?;
 
     log_event(
-        EventType::PolicyUpdated as i32,
+        EventType::PolicyUpdated,
         policy.uuid.as_ref(),
         &org_id,
         &headers.user.uuid,
@@ -2994,7 +2994,7 @@ async fn revoke_member_impl(
             member.save(conn).await?;
 
             log_event(
-                EventType::OrganizationUserRevoked as i32,
+                EventType::OrganizationUserRevoked,
                 &member.uuid,
                 org_id,
                 &headers.user.uuid,
@@ -3092,7 +3092,7 @@ async fn restore_member_impl(
             member.save(conn).await?;
 
             log_event(
-                EventType::OrganizationUserRestored as i32,
+                EventType::OrganizationUserRestored,
                 &member.uuid,
                 org_id,
                 &headers.user.uuid,
@@ -3294,7 +3294,7 @@ async fn post_groups(
     let group = group_request.to_group(&org_id);
 
     log_event(
-        EventType::GroupCreated as i32,
+        EventType::GroupCreated,
         &group.uuid,
         &org_id,
         &headers.user.uuid,
@@ -3417,7 +3417,7 @@ async fn put_group(
     // change the caller may not make, and an event written before it recorded an update that never
     // happened.
     log_event(
-        EventType::GroupUpdated as i32,
+        EventType::GroupUpdated,
         &group_id,
         &org_id,
         &headers.user.uuid,
@@ -3766,7 +3766,7 @@ async fn add_update_group(
             user_entry.save(conn).await?;
 
             log_event(
-                EventType::OrganizationUserUpdatedGroups as i32,
+                EventType::OrganizationUserUpdatedGroups,
                 &assigned_member,
                 &org_id,
                 &headers.user.uuid,
@@ -3890,7 +3890,7 @@ async fn delete_authorized_group(
     conn: &DbConn,
 ) -> EmptyResult {
     log_event(
-        EventType::GroupDeleted as i32,
+        EventType::GroupDeleted,
         &group.uuid,
         org_id,
         &headers.user.uuid,
@@ -4047,7 +4047,7 @@ async fn put_group_members(
         user_entry.save(&conn).await?;
 
         log_event(
-            EventType::OrganizationUserUpdatedGroups as i32,
+            EventType::OrganizationUserUpdatedGroups,
             &assigned_member,
             &org_id,
             &headers.user.uuid,
@@ -4099,7 +4099,7 @@ async fn post_delete_group_member(
     }
 
     log_event(
-        EventType::OrganizationUserUpdatedGroups as i32,
+        EventType::OrganizationUserUpdatedGroups,
         &member_id,
         &org_id,
         &headers.user.uuid,
@@ -4235,7 +4235,7 @@ async fn recover_account(
     nt.send_logout(&user, None, &conn).await;
 
     log_event(
-        EventType::OrganizationUserAdminResetPassword as i32,
+        EventType::OrganizationUserAdminResetPassword,
         &member_id,
         &org_id,
         &headers.user.uuid,
@@ -4362,9 +4362,9 @@ async fn put_reset_password_enrollment(
     membership.save(&conn).await?;
 
     let event_type = if membership.reset_password_key.is_some() {
-        EventType::OrganizationUserResetPasswordEnroll as i32
+        EventType::OrganizationUserResetPasswordEnroll
     } else {
-        EventType::OrganizationUserResetPasswordWithdraw as i32
+        EventType::OrganizationUserResetPasswordWithdraw
     };
 
     log_event(event_type, &membership.uuid, &org_id, &headers.user.uuid, headers.device.atype, &headers.ip.ip, &conn)
